Refactor exceptions in the project
This server has a lot of ") catch(...) (" blocks and
this was bad. After that i decided to remove all
exceptions from handlers and make boost errors insted.
But another problem was with boost JSON. This library doesn't
support non-exception boost::value methods.

So I refactored exceptions insted of removing them.

- https://www.boost.org/community/error_handling.html