Add DataBase to a server
In this commit I added a postgres database to a server.

In the config dog retiement time was added. After that time dog
leaves a game and it's score and playtime transfer to a
database. After the game players can see the best players.

To add this ability to the server the followed classes were
created:
	UseCases - interface to a common db adctions
	DogRepository - interface to a database on a model level
	DogRepositoryImpl - realization of the DogRepository class
	DataBase - store of all the Repositories

	ConnectionPool - multithreading class that operates with
		connetions to a database and makes it thread safety